rename: 상세페이지 api 폴더 구조 변경
